requirements.txt
requests==2.32.3
ib-insync==0.9.86
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
PyYAML==6.0.1
